Fix interpolation of ivec2
Previously interpolation of ivec2 would result in the last element being
smaller than the specified end element. This was due to truncation of
the non-integer part of the float in vec2 to ivec2 casting. To remedy
this we have replaced the implicit casting with rounding which mitigates
the problem.

We have also replaced the iterative interpolation with an explicit
calculation each time to aid with accuracy, at the cost of speed.