Populate vertex normals
This commit implement vertex normal construction, we find all triangles
sharing a vertex then calculate that vertex's normal by taking the
average of the face normals of the triangles.

This is a fairly arbitrary implementation that can later be extended to
factor size of faces etc into account to weight the vertex normal.