Split and duplcate projects into basic and extended forms
The rasteriser is at the 50% completion mark, so `rasteriser` should now
be left untouched, any extensions should be done to `raster-extensions`.

The raytracer is currently with extensions for both `raytracer` and
`raytracer-extensions`, the next step is to strip `raytracer` of our
extensions so it is back to the 50% completion mark. All further
extension should be applied to `raytracer-extensions`.